refactor #244: move kibble reward to utils
