+++ utils/getKibbleRewardMessage/__tests__/index.ts
@@ -1,0 +1,8 @@
+import { getKibbleRewardMessage } from '..'
+
+describe('getKibbleRewardMessage', () => {
+  it('returns a string as a reward', () => {
+    const result = getKibbleRewardMessage()
+    expect(typeof result).toBe('string')
+  })
+})
